grammatiker/ (saulius@koala.ibt.lt) Checking download of the Grammatica archive -- making installation of this dependency idempotent.
git-svn-id: svn+ssh://saulius-grazulis.lt/home/saulius/svn-repositories/grammatiker/trunk@107 8a4994de-5b18-4c12-b7b7-b390355806d6

+++ dependencies/LinuxMint-17.1/install.sh
@@ -9,11 +9,17 @@ GRAMMATICA_SRV=http://grammatica.percederberg.net
 GRAMMATICA_URL=${GRAMMATICA_SRV}/download/stable/1.6/grammatica-1.6.zip
 GRAMMATICA_ZIP=$(basename ${GRAMMATICA_URL})
 
-wget -O ${TOOL_DIR}/${GRAMMATICA_ZIP} ${GRAMMATICA_URL}
+if [ ! -f ${TOOL_DIR}/${GRAMMATICA_ZIP} ]
+then
+    wget -O ${TOOL_DIR}/${GRAMMATICA_ZIP} ${GRAMMATICA_URL}
 
-(
-    cd ${TOOL_DIR}
-    unzip ${GRAMMATICA_ZIP}
-    cd $(basename ${GRAMMATICA_ZIP} .zip)
-    ant -k
-)
+    (
+        cd ${TOOL_DIR}
+        unzip ${GRAMMATICA_ZIP}
+        cd $(basename ${GRAMMATICA_ZIP} .zip)
+        ant -k
+    )
+else
+    echo "$0: ${TOOL_DIR}/${GRAMMATICA_ZIP} is already downloaded -" \
+        "will not replace it"
+fi
